+++ BondageClub/Screens/Inventory/ItemHands/SpankingToys/SpankingToy.js
@@ -1,177 +1,158 @@
 "use strict";
-const SpankingInventory = [
+const InventoryItemHandsSpankingToysOptions = [
 	{
 		Name: "Crop",
+		Property: { Type: null },
 		Bonus: [{ Type: "KidnapDomination", Factor: 3 }],
 		ExpressionTrigger: [{ Group: "Blush", Name: "Low", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }]
 	}, {
 		Name: "Flogger",
+		Property: { Type: "Flogger" },
 		Bonus: [{ Type: "KidnapDomination", Factor: 3 }],
 		ExpressionTrigger: [{ Group: "Blush", Name: "Low", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }]
 	}, {
 		Name: "Cane",
+		Property: { Type: "Cane" },
 		Bonus: [{ Type: "KidnapDomination", Factor: 3 }],
 		ExpressionTrigger: [{ Group: "Blush", Name: "Medium", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }, { Group: "Eyes", Name: "Wink", Timer: 5 }]
 	}, {
 		Name: "HeartCrop",
+		Property: { Type: "HeartCrop" },
 		Bonus: [{ Type: "KidnapDomination", Factor: 3 }],
 		ExpressionTrigger: [{ Group: "Blush", Name: "Medium", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }]
 	}, {
 		Name: "Paddle",
+		Property: { Type: "Paddle" },
 		Bonus: [{ Type: "KidnapDomination", Factor: 3 }],
 		ExpressionTrigger: [{ Group: "Blush", Name: "High", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }, { Group: "Eyes", Name: "Closed", Timer: 5 }]
 	}, {
 		Name: "WhipPaddle",
+		Property: { Type: "WhipPaddle" },
 		Bonus: [{ Type: "KidnapDomination", Factor: 3 }],
 		ExpressionTrigger: [{ Group: "Blush", Name: "Medium", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }, { Group: "Eyes", Name: "Wink", Timer: 5 }]
 	}, {
 		Name: "Whip",
+		Property: { Type: "Whip" },
 		Bonus: [{ Type: "KidnapDomination", Factor: 3 }],
 		ExpressionTrigger: [{ Group: "Blush", Name: "Medium", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }, { Group: "Eyes", Name: "Wink", Timer: 5 }]
 	}, {
 		Name: "CattleProd",
+		Property: { Type: "CattleProd" },
 		Bonus: [{ Type: "KidnapDomination", Factor: 3 }],
 		ExpressionTrigger: [{ Group: "Blush", Name: "Medium", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }, { Group: "Eyes", Name: "Wink", Timer: 5 }]
 	}, {
 		Name: "TennisRacket",
+		Property: { Type: "TennisRacket" },
 		ExpressionTrigger: [{ Group: "Blush", Name: "Low", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }]
 	}, {
 		Name: "Feather",
+		Property: { Type: "Feather" },
 		ExpressionTrigger: [{ Group: "Blush", Name: "Medium", Timer: 10 }, {Group: "Eyes", Name: "Closed", Timer: 10}, { Group: "Mouth", Name: "Grin", Timer: 10}, { Group: "Eyebrows", Name: "Soft", Timer: 10}]
 	}, {
 		Name: "FeatherDuster",
+		Property: { Type: "FeatherDuster" },
 		ExpressionTrigger: [{ Group: "Blush", Name: "High", Timer: 10 }, { Group: "Eyes", Name: "Closed", Timer: 10}, { Group: "Mouth", Name: "Laughing", Timer: 10}, { Group: "Eyebrows", Name: "Soft", Timer: 10}]
 	}, {
 		Name: "IceCube",
+		Property: { Type: "IceCube" },
 		ExpressionTrigger: [{ Group: "Blush", Name: "Medium", Timer: 10}, { Group: "Mouth", Name: "Angry", Timer: 10}, { Group: "Eyes", Name: "Daydream", Timer: 10}, { Group: "Eyebrows", Name: "Soft", Timer: 10}]
 	}, {
 		Name: "WartenbergWheel",
+		Property: { Type: "WartenbergWheel" },
 		ExpressionTrigger: [{ Group: "Blush", Name: "Medium", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }, { Group: "Mouth", Name: "Frown", Timer: 10}, { Group: "Eyes", Name: "Daydream", Timer: 10}]
 	}, {
 		Name: "VibratingWand",
+		Property: { Type: "VibratingWand" },
 		ExpressionTrigger: [{ Group: "Blush", Name: "High", Timer: 10 }, { Group: "Eyes", Name: "VeryLewd", Timer: 5 }, { Group: "Eyebrows", Name: "Soft", Timer: 10}, { Group: "Mouth", Name: "Open", Timer: 10}]
 	}, {
 		Name: "SmallVibratingWand",
+		Property: { Type: "SmallVibratingWand" },
 		ExpressionTrigger: [{ Group: "Blush", Name: "Medium", Timer: 10 }, { Group: "Eyes", Name: "Dazed", Timer: 5 }, { Group: "Eyebrows", Name: "Soft", Timer: 10}, { Group: "Mouth", Name: "HalfOpen", Timer: 10}]
 	}, {
 		Name: "CandleWax",
+		Property: { Type: "CandleWax" },
 		ExpressionTrigger: [{ Group: "Blush", Name: "Medium", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }, { Group: "Eyes", Name: "Surprised", Timer: 10}, { Group: "Mouth", Name: "Frown", Timer: 10}]
 	}, {
 		Name: "LargeDildo",
+		Property: { Type: "LargeDildo" },
 		ExpressionTrigger: [{ Group: "Blush", Name: "Medium", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }, { Group: "Mouth", Name: "HalfOpen", Timer: 10}, { Group: "Eyes", Name: "Lewd", Timer: 10}]
 	}, {
 		Name: "PetToy",
+		Property: { Type: "PetToy" },
 		ExpressionTrigger: [{ Group: "Blush", Name: "Low", Timer: 10 }, { Group: "Eyes", Name: "WinkL", Timer: 10 }, { Group: "Mouth", Name: "Smirk", Timer: 10 }]
 	},{
 		Name: "Vibrator",
+		Property: { Type: "Vibrator" },
 		ExpressionTrigger: [{ Group: "Blush", Name: "High", Timer: 10 }, { Group: "Eyes", Name: "VeryLewd", Timer: 10 }, { Group: "Mouth", Name: "Open", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10}]
 	},{
 		Name: "Belt",
+		Property: { Type: "Belt" },
 		ExpressionTrigger: [{ Group: "Blush", Name: "Low", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }, { Group: "Mouth", Name: "Frown", Timer: 10}]
 	}, {
 		Name: "Hairbrush",
+		Property: { Type: "Hairbrush" },
 		ExpressionTrigger: [{ Group: "Blush", Name: "Low", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }, { Group: "Mouth", Name: "Frown", Timer: 10}]
 	}, {
 		Name: "SmallDildo",
+		Property: { Type: "SmallDildo" },
 		ExpressionTrigger: [{ Group: "Blush", Name: "Low", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }, { Group: "Mouth", Name: "Frown", Timer: 10}, { Group: "Eyes", Name: "Lewd", Timer: 5 }, ]
 	}, {
 		Name: "ElectricToothbrush",
+		Property: { Type: "ElectricToothbrush" },
 		ExpressionTrigger: [{ Group: "Blush", Name: "Medium", Timer: 10 }, {Group: "Eyes", Name: "Closed", Timer: 10}, { Group: "Mouth", Name: "Grin", Timer: 10}, { Group: "Eyebrows", Name: "Soft", Timer: 10}]
 	}, {
 		Name: "Toothbrush",
+		Property: { Type: "Toothbrush" },
 		ExpressionTrigger: [{ Group: "Blush", Name: "Medium", Timer: 10 }, {Group: "Eyes", Name: "Closed", Timer: 10}, { Group: "Mouth", Name: "Grin", Timer: 10}, { Group: "Eyebrows", Name: "Soft", Timer: 10}]
 	}, {
 		Name: "ShockWand",
+		Property: { Type: "ShockWand" },
 		ExpressionTrigger: [{ Group: "Blush", Name: "Medium", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }, { Group: "Eyes", Name: "Wink", Timer: 5 }]
-	},{
-	  Name: "Lotion",
-	  ExpressionTrigger: [{ Group: "Blush", Name: "Low", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }, { Group: "Mouth", Name: "Frown", Timer: 10}]
+	}, {
+		Name: "Lotion",
+		Property: { Type: "Lotion" },
+		ExpressionTrigger: [{ Group: "Blush", Name: "Low", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }, { Group: "Mouth", Name: "Frown", Timer: 10 }]
+	}, {
+		Name: "Ruler",
+		Property: { Type: "Ruler" },
+		ExpressionTrigger: [{ Group: "Blush", Name: "High", Timer: 10 }, { Group: "Eyebrows", Name: "Soft", Timer: 10 }, { Group: "Eyes", Name: "Closed", Timer: 5 }]
 	},
-	
 ];
-
-var SpankingInventoryOffset = 0;
-var SpankingNextButton = false;
-var SpankingPlayerInventory;
 
 // Loads the item extension properties
 function InventoryItemHandsSpankingToysLoad() {
-	SpankingPlayerInventory = SpankingInventory.filter(x => Player.Inventory.map(i => i.Name).includes("SpankingToys" + x.Name));
-	if (DialogFocusItem.Property == null) {
-		DialogFocusItem.Property = { Type: "Crop" };
-		CharacterRefresh(CharacterGetCurrent(), false);
-	}
-	if (SpankingPlayerInventory.length > 6) SpankingNextButton = true;
+	ExtendedItemLoad(InventorySpankingToysOwnedToys(), "SelectSpankingToysType");
 }
 
-// Item groups that called the function (just forwarding it to SpankingToysDraw()) *brute force, might be a better way to do this
+// Draw the item extension screen
 function InventoryItemHandsSpankingToysDraw() {
-
-	// Draw the header and item
-	if (SpankingNextButton) DrawButton(1775, 25, 90, 90, "", "White", "Icons/Next.png");
-	DrawRect(1387, 55, 225, 275, "white");
-	DrawImageResize("Assets/" + DialogFocusItem.Asset.Group.Family + "/" + DialogFocusItem.Asset.Group.Name + "/Preview/" + DialogFocusItem.Asset.Name + ".png", 1389, 57, 221, 221);
-	DrawTextFit(DialogFocusItem.Asset.Description, 1500, 310, 221, "black");
-	DrawText(DialogFind(Player, "SelectSpankingToysType"), 1500, 375, "white", "gray");
-
-	// Draw the buttons 6 at a time, in a 2x3 grid
-	for (let I = SpankingInventoryOffset; (I < SpankingPlayerInventory.length) && (I < SpankingInventoryOffset + 6); I++) {
-		var offset = I - SpankingInventoryOffset;
-		var X = 1080 + (offset % 3 * 305);
-		var Y = 430 + (Math.floor(offset / 3) * 300);
-
-		DrawButton(X, Y, 225, 225, "", ((DialogFocusItem.Property.Type == SpankingPlayerInventory[I].Name) ? "#888888" : "White"));
-		DrawImage("Screens/Inventory/" + DialogFocusItem.Asset.Group.Name + "/" + DialogFocusItem.Asset.Name + "/" + SpankingPlayerInventory[I].Name + ".png", X, Y);
-		DrawText(DialogFind(Player, "SpankingToysType" + SpankingPlayerInventory[I].Name), X + 115, Y + 250, "white", "gray");
-	};
+	ExtendedItemDraw(InventorySpankingToysOwnedToys(), "SpankingToysType");
 }
 
 // Catches the item extension clicks
 function InventoryItemHandsSpankingToysClick() {
-
-	// Menu buttons
-	if ((MouseX >= 1885) && (MouseX <= 1975) && (MouseY >= 25) && (MouseY <= 110)) DialogFocusItem = null;
-	if ((MouseX >= 1775) && (MouseX <= 1865) && (MouseY >= 25) && (MouseY <= 110) && (SpankingNextButton)) SpankingInventoryOffset += 6;
-	if (SpankingInventoryOffset >= SpankingPlayerInventory.length) SpankingInventoryOffset = 0;
-
-	// Item buttons
-	for (let I = SpankingInventoryOffset; (I < SpankingPlayerInventory.length) && (I < SpankingInventoryOffset + 6); I++) {
-		var nextItem = SpankingPlayerInventory[I].Name;
-		var offset = I - SpankingInventoryOffset;
-		var X = 1080 + (offset % 3 * 305);
-		var Y = 430 + (Math.floor(offset / 3) * 300);
-
-		if ((MouseX >= X) && (MouseX <= X + 225) && (MouseY >= Y) && (MouseY <= Y + 225) && (DialogFocusItem.Property.Type != nextItem))
-			InventorySpankingToySetType(nextItem);
-	}
+	ExtendedItemClick(InventorySpankingToysOwnedToys());
 }
 
-// Uses spanking toy type (cane, crop, flogger, etc.)
-function InventorySpankingToySetType(NewType) {
-
-	// Sets the type
-	var C = (Player.FocusGroup != null) ? Player : CurrentCharacter;
-	if (CurrentScreen == "ChatRoom") {
-		DialogFocusItem = InventoryGet(C, C.FocusGroup.Name);
-		InventoryItemHandsSpankingToysLoad();
-	}
-	DialogFocusItem.Property.Type = NewType;
-	CharacterRefresh(C);
-
-	// Prepares the chat message to be published
+function InventoryItemHandsSpankingToysPublishAction(C, Option) {
 	var msg = C.ID == 0 ? "SpankingToysSetPlayer" : "SpankingToysSetOthers";
 	var Dictionary = [];
-	Dictionary.push({Tag: "SourceCharacter", Text: Player.Name, MemberNumber: Player.MemberNumber});
-	Dictionary.push({Tag: "TargetCharacter", Text: C.Name, MemberNumber: C.MemberNumber});
-	Dictionary.push({Tag: "ItemUsed", AssetName : "SpankingToys" + ((NewType) ? NewType : "Crop")});
+	Dictionary.push({ Tag: "SourceCharacter", Text: Player.Name, MemberNumber: Player.MemberNumber });
+	Dictionary.push({ Tag: "TargetCharacter", Text: C.Name, MemberNumber: C.MemberNumber });
+	Dictionary.push({ Tag: "ItemUsed", AssetName: "SpankingToys" + Option.Name });
 	ChatRoomPublishCustomAction(msg, true, Dictionary);
+}
 
-	// Exit from item when done
-	if (DialogInventory != null) {
-		DialogFocusItem = null;
-		DialogMenuButtonBuild(C);
-	}
+function InventoryItemHandsSpankingToysNpcDialog(C, Option) {
+	C.CurrentDialog = DialogFind(C, "SpankingToys" + Option.Name, "ItemHands");
+}
 
+/**
+ * Returns a list of the spanking toys that the player owns
+ * @returns {ExtendedItemOption[]} The subset of SpankingToys options the player can select from
+ */
+function InventorySpankingToysOwnedToys() {
+	return InventoryItemHandsSpankingToysOptions.filter(x => Player.Inventory.map(i => i.Name).includes("SpankingToys" + x.Name));
 }
 
 // Get the type of spanking toy that the character is holding